LPD-89542 Persist selected translation language across preview open/close
